Add tutorial flow and focus overlays
This commit adds a new tutorial intro screen and a set of animated focus overlays for gameplay elements. It updates the tutorial scene flow, adds localized tutorial strings, and adjusts the quick menu labels

+++ src/quirkle/engine/Entity.java
@@ -46,8 +46,8 @@ public abstract class Entity {
 
     public double x = 0;
     public double y = 0;
-    public int targetX = 0;
-    public int targetY = 0;
+    public double targetX = 0;
+    public double targetY = 0;
     public double rotation = 0;
 
     /**
